backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-cloud-storage==2.10.0
google-cloud-speech>=2.20.0,<3.0.0
google-cloud-documentai==2.20.1
google-api-python-client==2.100.0
google-auth>=2.45.0,<3.0.0
google-auth[requests]>=2.45.0,<3.0.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.3.5
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
pydub==0.25.1
importlib-metadata>=4.6.0

